# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Video Intelligence",
        "title": "Video Intelligence",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "beijingyingnuoweixun",
            "url": "https://e.gitee.com/beijingyingnuoweixun"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "localhost:8090",
    "basePath": "/api/v1",
    "paths": {
        "/devices/{id}/channels": {
            "post": {
                "description": "通过此接口在设备下新增通道，获取通道id",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "通道新增接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "设备id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "通道备注",
                        "name": "memo",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "播放类型，pull 媒体服务器拉流，push 摄像头推流,默认push",
                        "name": "streamtype",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "静态拉流地址，streamtype=pull 时生效。",
                        "name": "url",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "1000": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "1001": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "1002": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "1003": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}